[workshops] do not allow user to delete workshop that has offerings

+++ app/Livewire/Workshops/DeleteWorkshopWarning.php
@@ -4,6 +4,8 @@ namespace App\Livewire\Workshops;
 
 use App\Models\Workshop;
 // use Illuminate\Support\Facades\Gate;
+
+use Illuminate\Database\QueryException;
 use LivewireUI\Modal\ModalComponent;
 use App\Traits\ToastNotifications;
 
@@ -17,7 +19,7 @@ class DeleteWorkshopWarning extends ModalComponent
 
     public function mount()
     {
-        $this->workshop = Workshop::find($this->id);
+        $this->workshop = Workshop::with("offerings")->find($this->id);
         $this->title = $this->workshop->title_ar;
         if (app()->getLocale() == "en") {
             $this->title = $this->workshop->title_en;
@@ -26,17 +28,27 @@ class DeleteWorkshopWarning extends ModalComponent
 
     public function delete()
     {
-        // TODO: create a policy first
-        // if (Gate::denies("delete", $this->workshop)) {
-        //     $this->toastError(__("messages.not_auth_for_operation"));
-        //     $this->dispatch("closeModal");
-        // }
+        try {
+            // TODO: create a policy first
+            // if (Gate::denies("delete", $this->workshop)) {
+            //     $this->toastError(__("messages.not_auth_for_operation"));
+            //     $this->dispatch("closeModal");
+            // }
 
-        // TODO: must handle when workshops has offerings
-        $this->workshop->delete();
-        $this->dispatch("workshop-deleted");
-        $this->toastSuccess(__("messages.deleted_successfully"));
-        $this->closeModal();
+            // TODO: must handle when workshops has offerings
+            if (count($this->workshop->offerings)) {
+                $this->toastError(__("messages.workshop_has_offerings"));
+                $this->closeModal();
+                return;
+            }
+            $this->workshop->delete();
+            $this->dispatch("workshop-deleted");
+            $this->toastSuccess(__("messages.deleted_successfully"));
+            $this->closeModal();
+        } catch (QueryException $e) {
+            $this->toastError($e->getMessage());
+            $this->closeModal();
+        }
     }
 
     public function render()
